+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [Unreleased](https://codeberg.org/celema/console/compare/0.5.1...HEAD)
 
-No notable changes since the last release.
+### Added
+
+- Truecolor hex markup tags with a lowercase six-digit code: `<#ff7313>` and `<bg-#ff7313>`. Such tags previously passed through as unknown text; now they render (and broken pairs throw like any other tag).
 
 ## [0.5.1](https://codeberg.org/celema/console/src/tag/0.5.1) (2026-07-21)